refactor: move types to folder TYPES, correct imports and exports
